requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras
#
amqp==5.1.1
    # via kombu
argon2-cffi==23.1.0
    # via django
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
astroid==3.2.2
    # via pylint
attrs==22.1.0
    # via pytest
bandit==1.7.8
    # via -r requirements.in
billiard==4.2.0
    # via celery
boto3==1.24.78
    # via django-ses
botocore==1.27.78
    # via
    #   boto3
    #   s3transfer
build==1.2.1
    # via pip-tools
cachetools==5.3.3
    # via google-auth
celery==5.4.0
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via argon2-cffi-bindings
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coverage==6.4.4
    # via
    #   coverage
    #   pytest-cov
dill==0.3.6
    # via pylint
django==5.0.6
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   django-filter
    #   django-health-check
    #   django-model-utils
    #   django-phonenumber-field
    #   django-ses
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-yasg
django-cors-headers==4.3.1
    # via -r requirements.in
django-crum==0.7.9
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
django-health-check==3.18.2
    # via -r requirements.in
django-model-utils==4.5.1
    # via -r requirements.in
django-phonenumber-field==7.3.0
    # via
    #   -r requirements.in
    #   django-phonenumber-field
django-ses==4.1.0
    # via -r requirements.in
django-timezone-field==6.1.0
    # via -r requirements.in
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-nested-routers==0.94.1
    # via -r requirements.in
drf-yasg==1.21.7
    # via -r requirements.in
factory-boy==3.3.0
    # via -r requirements.in
faker==14.2.0
    # via factory-boy
freezegun==1.5.1
    # via -r requirements.in
google-api-core==2.19.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-translate
google-auth==2.30.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-translate
google-cloud-core==2.4.1
    # via google-cloud-translate
google-cloud-translate==3.15.3
    # via -r requirements.in
googleapis-common-protos==1.63.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.64.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.2
    # via google-api-core
gunicorn==22.0.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.3.7
    # via celery
markdown-it-py==2.2.0
    # via rich
mccabe==0.7.0
    # via pylint
mdurl==0.1.2
    # via markdown-it-py
packaging==21.3
    # via
    #   build
    #   drf-yasg
    #   gunicorn
    #   pytest
pathtools==0.1.2
    # via -r requirements.in
pbr==5.10.0
    # via stevedore
phonenumbers==8.13.17
    # via django-phonenumber-field
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==2.5.2
    # via pylint
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.31
    # via click-repl
proto-plus==1.23.0
    # via
    #   google-api-core
    #   google-cloud-translate
protobuf==4.25.3
    # via
    #   google-api-core
    #   google-cloud-translate
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psutil==5.9.8
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.22
    # via cffi
pygments==2.15.0
    # via rich
pyjwt==2.5.0
    # via djangorestframework-simplejwt
pylint==3.2.3
    # via
    #   -r requirements.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.5
    # via -r requirements.in
pylint-plugin-utils==0.8.2
    # via pylint-django
pyparsing==3.0.9
    # via packaging
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==7.2.0
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
pytest-cov==5.0.0
    # via -r requirements.in
pytest-django==4.8.0
    # via -r requirements.in
pytest-mock==3.14.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   celery
    #   faker
    #   freezegun
pytz==2022.2.1
    # via
    #   django-ses
    #   drf-yasg
pyyaml==6.0
    # via
    #   bandit
    #   drf-yasg
redis==5.0.5
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   google-api-core
rich==13.3.4
    # via bandit
rsa==4.9
    # via google-auth
s3transfer==0.6.0
    # via boto3
sentry-sdk==2.5.1
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.4
    # via django
stevedore==4.0.0
    # via bandit
tomlkit==0.11.4
    # via pylint
tzdata==2023.3
    # via celery
uritemplate==4.1.1
    # via drf-yasg
urllib3==1.26.18
    # via
    #   botocore
    #   requests
    #   sentry-sdk
uvicorn==0.30.1
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.38.1
    # via pip-tools
whitenoise==6.6.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
